+++ ExtendableEnums.Serialization.Newtonsoft/SerializableExtendableEnumDictionary.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+
+namespace ExtendableEnums.Serialization.Newtonsoft;
+
+/// <summary>
+/// A specialized dictionary to ensure proper serialization when an <see cref="ExtendableEnumBase{TEnumeration, TValue}" />
+/// type object is the key type.
+/// </summary>
+/// <typeparam name="TKey">The type of the key.</typeparam>
+/// <typeparam name="TValue">The type of the value.</typeparam>
+[JsonConverter(typeof(ExtendableEnumDictionaryJsonConverter))]
+public class SerializableExtendableEnumDictionary<TKey, TValue> : ExtendableEnumDictionary<TKey, TValue>
+        where TKey : IExtendableEnum
+{
+}