+++ ExtendableEnums/ExtendableEnumDictionary.cs
@@ -1,7 +1,5 @@
 ﻿using System.Diagnostics;
 using System.Runtime.Serialization;
-using ExtendableEnums.Serialization.Newtonsoft;
-using Newtonsoft.Json;
 
 namespace ExtendableEnums;
 
@@ -11,7 +9,6 @@ namespace ExtendableEnums;
 /// </summary>
 /// <typeparam name="TKey">The type of the key.</typeparam>
 /// <typeparam name="TValue">The type of the value.</typeparam>
-[JsonConverter(typeof(ExtendableEnumDictionaryJsonConverter))]
 [Serializable]
 public class ExtendableEnumDictionary<TKey, TValue> : Dictionary<TKey, TValue>
     where TKey : IExtendableEnum
@@ -117,4 +114,4 @@ public class ExtendableEnumDictionary<TKey, TValue> : Dictionary<TKey, TValue>
         Debug.WriteLine("Getting object data...");
         base.GetObjectData(info, context);
     }
-}
+}